Change to bigdecimal
kotlin querydsl serializer issue with Double to double

+++ src/main/kotlin/griffio/entity/Satellite.kt
@@ -1,8 +1,9 @@
 package griffio.entity
 
+import java.math.BigDecimal
 import javax.persistence.Entity
 
-@Entity class Satellite(var id : Long, var name : String, var diameter : Double) {
+@Entity class Satellite(var id: Long, var name: String, var diameter: BigDecimal) {
 
     override fun toString(): String {
         return "Satellite:$id:$name:$diameter (km)"
